- Changed the types of object hash indices in FArchive from size_t to DWORD.   This seems to fix crashes on GCC 64-bit builds when saving games. Not sure   if it was a GCC bug or my bug, since it worked fine with VC++, but since the   code that calculates the index only returns a DWORD, storing it as a size_t   was rather pointless.
SVN r969 (trunk)

+++ src/farchive.cpp
@@ -1459,12 +1459,12 @@ DWORD FArchive::HashObject (const DObject *obj) const
 
 DWORD FArchive::FindObjectIndex (const DObject *obj) const
 {
-	size_t index = m_ObjectHash[HashObject (obj)];
+	DWORD index = m_ObjectHash[HashObject (obj)];
 	while (index != TypeMap::NO_INDEX && m_ObjectMap[index].object != obj)
 	{
 		index = m_ObjectMap[index].hashNext;
 	}
-	return (DWORD)index;
+	return index;
 }
 
 void FArchive::UserWriteClass (const PClass *type)
